App working for a while, please help record video
