+++ MC2-17/DiaBeta/Feature/Pages/MainScreen/MainScreen.swift
@@ -6,7 +6,7 @@
 //
 
 import UIKit
-import UserNotifications
+//import UserNotifications
 
 var goodFoodInfo:[FoodInfo] = [] //collect good food data
 var badFoodInfo:[FoodInfo] = [] //collect bad food data
@@ -32,7 +32,7 @@ class MainScreen: UIViewController, UNUserNotificationCenterDelegate {
         super.viewDidLoad()
         self.topView.backgroundColor = UIColor(named: "AccentColor")
         bottomView.layer.cornerRadius = 10
-        registerLocal()
+//        registerLocal()
         newsView.layer.cornerRadius = 8
         newsView.layer.shadowOffset = CGSize(width: 5, height: 5)
         newsView.layer.shadowRadius = 5
@@ -56,15 +56,15 @@ class MainScreen: UIViewController, UNUserNotificationCenterDelegate {
   }
   
   @IBAction func seeAllGoodFood(_ sender: UIButton) {
-//    performSegue(withIdentifier: "toSeeAll", sender: self)
-//    print("All Good Food")
-    reminderLocal()
+    performSegue(withIdentifier: "toSeeAll", sender: self)
+    print("All Good Food")
+//    reminderLocal()
     
   }
   @IBAction func seeAllBadFood(_ sender: UIButton) {
-//    performSegue(withIdentifier: "toSeeAll", sender: self)
-//    print("All Bad Food")
-    reminderLocal()
+    performSegue(withIdentifier: "toSeeAll", sender: self)
+    print("All Bad Food")
+//    reminderLocal()
   }
   
 }
@@ -97,7 +97,7 @@ extension MainScreen: UICollectionViewDelegate, UICollectionViewDataSource{
     let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "foodCell", for: indexPath) as! GlucoseFoodCell
     
     if collectionView == foodCollectionView {
-      cell.foodImage.image = UIImage(data: (goodFoodInfo[indexPath.row].food?.photo)!as Data)
+      cell.foodImage.image = UIImage(data: (goodFoodInfo[indexPath.row].food?.photo)! as Data)
       cell.foodName.text = goodFoodInfo[indexPath.row].food?.name
       let stringArr = goodFoodInfo[indexPath.row].food?.category!
       let categoryArr = stringArr!.joined(separator: ",")
@@ -183,107 +183,107 @@ extension MainScreen {
 
 //MARK: - Notification Center
 
-extension MainScreen {
-  func registerLocal() {
-          let center = UNUserNotificationCenter.current()
-
-          center.requestAuthorization(options: [.alert, .badge, .sound]) { granted, error in
-              if granted {
-                 print("Yay!")
-              } else {
-                  print("D'oh!")
-              }
-          }
-
-      }
-  
-  func reminderLocal() {
-          registerCategories()
-
-          let center = UNUserNotificationCenter.current()
-          center.removeAllPendingNotificationRequests()
-
-          let content = UNMutableNotificationContent()
-          content.title = "Post-Glucose"
-          content.body = "Please input your post-meal glucose"
-          content.categoryIdentifier = "alarm"
-          content.userInfo = ["customData": "fizzbuzz"]
-          content.sound = .default
-
-          let trigger = UNTimeIntervalNotificationTrigger(timeInterval: 5, repeats: false)
-
-          let request = UNNotificationRequest(identifier: UUID().uuidString, content: content, trigger: trigger)
-          center.add(request)
-  }
-  
-  func scheduleLocal() {
-          registerCategories()
-
-          let center = UNUserNotificationCenter.current()
-          center.removeAllPendingNotificationRequests()
-
-          let content = UNMutableNotificationContent()
-          content.title = "Post-Glucose"
-          content.body = "Please input your post-meal glucose"
-          content.categoryIdentifier = "alarm"
-          content.userInfo = ["customData": "fizzbuzz"]
-          content.sound = .default
-
-      //ini buat set kapan notifnya muncul -> ini muncul tiap jam 10 malam
-            var dateComponents = DateComponents()
-            dateComponents.calendar = Calendar.current
-            dateComponents.hour = 22
-            dateComponents.minute = 00
-
-            let trigger = UNCalendarNotificationTrigger(dateMatching: dateComponents, repeats: true)
-
-            let request = UNNotificationRequest(identifier: UUID().uuidString, content: content, trigger: trigger)
-            center.add(request)
-  }
-  
-  func registerCategories() {
-      let center = UNUserNotificationCenter.current()
-      center.delegate = self
-
-      let show = UNNotificationAction(identifier: "show", title: "Tell me more...", options: .foreground)
-      let category = UNNotificationCategory(identifier: "alarm", actions: [show], intentIdentifiers: [])
-
-      center.setNotificationCategories([category])
-  }
-
-  func userNotificationCenter(_ center: UNUserNotificationCenter, didReceive response: UNNotificationResponse, withCompletionHandler completionHandler:
-                              @escaping() -> Void) {
-    
-    let storyboard = UIStoryboard(name: "EditMeals", bundle: nil) //change to edit food with custom data
-    
-    //initiate the view controller from storyboard
-    let editFoodVC = storyboard.instantiateViewController(withIdentifier: "EditMealsViewController")
-
-    self.present(editFoodVC, animated: true)
-        
-//      let userInfo = response.notification.request.content.userInfo
-      // Print message ID
-    
-//      if let customData = userInfo["customData"] as? String {
+//extension MainScreen {
+//  func registerLocal() {
+//          let center = UNUserNotificationCenter.current()
 //
-//          print("Custom data received: \(customData)") //MASUKIN DATE TIME NYA DISINI
-//
-//          switch response.actionIdentifier {
-//          case UNNotificationDefaultActionIdentifier:
-//              //buat user swipe pas unlock
-//                  print("Default identifier")
-//
-//          case "show" :
-//              print("Show more information...")
-//
-//          default:
-//              break
+//          center.requestAuthorization(options: [.alert, .badge, .sound]) { granted, error in
+//              if granted {
+//                 print("Yay!")
+//              } else {
+//                  print("D'oh!")
+//              }
 //          }
+//
 //      }
-
-      completionHandler()
-  }
-  
-}
+//  
+//  func reminderLocal() {
+//          registerCategories()
+//
+//          let center = UNUserNotificationCenter.current()
+//          center.removeAllPendingNotificationRequests()
+//
+//          let content = UNMutableNotificationContent()
+//          content.title = "Post-Glucose"
+//          content.body = "Please input your post-meal glucose"
+//          content.categoryIdentifier = "alarm"
+//          content.userInfo = ["customData": "fizzbuzz"]
+//          content.sound = .default
+//
+//          let trigger = UNTimeIntervalNotificationTrigger(timeInterval: 5, repeats: false)
+//
+//          let request = UNNotificationRequest(identifier: UUID().uuidString, content: content, trigger: trigger)
+//          center.add(request)
+//  }
+//  
+//  func scheduleLocal() {
+//          registerCategories()
+//
+//          let center = UNUserNotificationCenter.current()
+//          center.removeAllPendingNotificationRequests()
+//
+//          let content = UNMutableNotificationContent()
+//          content.title = "Post-Glucose"
+//          content.body = "Please input your post-meal glucose"
+//          content.categoryIdentifier = "alarm"
+//          content.userInfo = ["customData": "fizzbuzz"]
+//          content.sound = .default
+//
+//      //ini buat set kapan notifnya muncul -> ini muncul tiap jam 10 malam
+//            var dateComponents = DateComponents()
+//            dateComponents.calendar = Calendar.current
+//            dateComponents.hour = 22
+//            dateComponents.minute = 00
+//
+//            let trigger = UNCalendarNotificationTrigger(dateMatching: dateComponents, repeats: true)
+//
+//            let request = UNNotificationRequest(identifier: UUID().uuidString, content: content, trigger: trigger)
+//            center.add(request)
+//  }
+//  
+//  func registerCategories() {
+//      let center = UNUserNotificationCenter.current()
+//      center.delegate = self
+//
+//      let show = UNNotificationAction(identifier: "show", title: "Tell me more...", options: .foreground)
+//      let category = UNNotificationCategory(identifier: "alarm", actions: [show], intentIdentifiers: [])
+//
+//      center.setNotificationCategories([category])
+//  }
+//
+//  func userNotificationCenter(_ center: UNUserNotificationCenter, didReceive response: UNNotificationResponse, withCompletionHandler completionHandler:
+//                              @escaping() -> Void) {
+//    
+//    let storyboard = UIStoryboard(name: "EditMeals", bundle: nil) //change to edit food with custom data
+//    
+//    //initiate the view controller from storyboard
+//    let editFoodVC = storyboard.instantiateViewController(withIdentifier: "EditMealsViewController")
+//
+//    self.present(editFoodVC, animated: true)
+//        
+////      let userInfo = response.notification.request.content.userInfo
+//      // Print message ID
+//    
+////      if let customData = userInfo["customData"] as? String {
+////
+////          print("Custom data received: \(customData)") //MASUKIN DATE TIME NYA DISINI
+////
+////          switch response.actionIdentifier {
+////          case UNNotificationDefaultActionIdentifier:
+////              //buat user swipe pas unlock
+////                  print("Default identifier")
+////
+////          case "show" :
+////              print("Show more information...")
+////
+////          default:
+////              break
+////          }
+////      }
+//
+//      completionHandler()
+//  }
+//  
+//}
