+++ MC2-17/DiaBeta/Feature/Pages/Meals/MealTableViewCell.xib
@@ -27,7 +27,7 @@
                             <color key="baseForegroundColor" red="0.7843137255" green="0.7843137255" blue="0.79215686269999996" alpha="1" colorSpace="custom" customColorSpace="displayP3"/>
                         </buttonConfiguration>
                     </button>
-                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="4vW-Tf-cpb">
+                    <imageView autoresizesSubviews="NO" clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="4vW-Tf-cpb">
                         <rect key="frame" x="17.5" y="7" width="84" height="83"/>
                         <constraints>
                             <constraint firstAttribute="width" constant="84" id="0Kf-EC-aqK"/>
