+++ MC2-17/DiaBeta/Feature/Pages/Meals/MealViewController.swift
@@ -92,7 +92,11 @@ class MealViewController: UIViewController, UITableViewDelegate, UITableViewData
      //If you want pass data while segue you can use prepare segue method
     }
     
-    
+  @IBAction func unwindToMeal(_ unwindSegue: UIStoryboardSegue) {
+    let sourceViewController = unwindSegue.source
+    // Use data from the view controller which initiated the unwind segue
+  }
+
   
     func getWeeks() {
         let dateCur = Date()
@@ -150,6 +154,7 @@ class MealViewController: UIViewController, UITableViewDelegate, UITableViewData
 //        cell.myImageView.image = UIImage(data: foodInfos[indexPath.row].photo !as Data
         cell.myImageView.image = UIImage(data: (foodInfos[indexPath.row].photo)!as Data)
       cell.myImageView.contentMode = .scaleAspectFit
+      cell.myImageView.layer.masksToBounds = true
     
         return cell
     }
